+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 Please see [CONTRIBUTING.md](./CONTRIBUTING.md) on how to contribute to Cucumber.
 
 ## [Unreleased]
+- Add error message for pending steps ([#2392](https://github.com/cucumber/cucumber-js/pull/2393))
 
 ## [10.4.0] - 2024-04-07
 ### Added